Database
Contains information for the database, .sql for creating tables and importing data. will add an cvs later :) 
if you have mysql workbench, you can use the "db" folders to open it else you would have to use the .sql to create the tables and the .sql in the "datafolder" to import the data. later when i add cvs, ill seprate the .sql and .csv into different folder in the "datafolder" but for now its just .sql 

PS. i also added an "readme" for reminder